+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## dbt-teradata 0.19.0a
 
 ### Features
+* Added support for CREATE TABLE parameters (https://github.com/Teradata/dbt-teradata/issues/2)
 
 ### Fixes